Exec Report Builder first pass.
* Minimal implementation of builder and state.
* Refactor existing functions, move to new package.
* Update selectReport to use builder.
* Tests not passing.
