--- execute/errors.go
@@ -1,5 +1,0 @@
-package execute
-
-import "errors"
-
-var errEmptyReport = errors.New("no messages can fit in the report")
