+++ execute/plugin.go
@@ -1,10 +1,7 @@
 package execute
 
 import (
-	"bytes"
 	"context"
-	"encoding/hex"
-	"errors"
 	"fmt"
 	"sort"
 	"sync/atomic"
@@ -19,7 +16,8 @@ import (
 	"github.com/smartcontractkit/libocr/offchainreporting2plus/types"
 	libocrtypes "github.com/smartcontractkit/libocr/ragep2p/types"
 
-	"github.com/smartcontractkit/chainlink-ccip/internal/libs/slicelib"
+	"github.com/smartcontractkit/chainlink-ccip/execute/report"
+	"github.com/smartcontractkit/chainlink-ccip/execute/temp"
 	"github.com/smartcontractkit/chainlink-ccip/internal/reader"
 	"github.com/smartcontractkit/chainlink-ccip/pluginconfig"
 	"github.com/smartcontractkit/chainlink-ccip/plugintypes"
@@ -40,7 +38,7 @@ type Plugin struct {
 	homeChain   reader.HomeChain
 
 	oracleIDToP2pID map[commontypes.OracleID]libocrtypes.PeerID
-	tokenDataReader TokenDataReader
+	tokenDataReader temp.TokenDataReader
 	lastReportTS    *atomic.Int64
 	lggr            logger.Logger
 }
@@ -53,7 +51,7 @@ func NewPlugin(
 	reportCodec cciptypes.ExecutePluginCodec,
 	msgHasher cciptypes.MessageHasher,
 	homeChain reader.HomeChain,
-	tokenDataReader TokenDataReader,
+	tokenDataReader temp.TokenDataReader,
 	lggr logger.Logger,
 ) *Plugin {
 	lastReportTS := &atomic.Int64{}
@@ -245,206 +243,6 @@ func (p *Plugin) ObservationQuorum(outctx ocr3types.OutcomeContext, query types.
 	return ocr3types.QuorumFPlusOne, nil
 }
 
-// TokenDataReader is an interface for reading extra token data from an async process.
-// TODO: Build a token data reading process.
-//
-//go:generate mockery --quiet --name TokenDataReader --output ../internal/mocks --case=underscore
-type TokenDataReader interface {
-	ReadTokenData(ctx context.Context, srcChain cciptypes.ChainSelector, num cciptypes.SeqNum) ([][]byte, error)
-}
-
-// buildSingleChainReportMaxSize generates the largest report which fits into maxSizeBytes.
-// See buildSingleChainReport for more details about how a report is built.
-func buildSingleChainReportMaxSize(
-	ctx context.Context,
-	lggr logger.Logger,
-	hasher cciptypes.MessageHasher,
-	tokenDataReader TokenDataReader,
-	encoder cciptypes.ExecutePluginCodec,
-	report plugintypes.ExecutePluginCommitDataWithMessages,
-	maxSizeBytes int,
-) (cciptypes.ExecutePluginReportSingleChain, int, plugintypes.ExecutePluginCommitDataWithMessages, error) {
-	finalReport, encodedSize, _, err :=
-		buildSingleChainReport(ctx, lggr, hasher, tokenDataReader, encoder, report, nil)
-	if err != nil {
-		return cciptypes.ExecutePluginReportSingleChain{},
-			0,
-			plugintypes.ExecutePluginCommitDataWithMessages{},
-			fmt.Errorf("unable to build a single chain report (max): %w", err)
-	}
-
-	// return fully executed report
-	if encodedSize <= maxSizeBytes {
-		report = markNewMessagesExecuted(finalReport, report)
-		return finalReport, encodedSize, report, nil
-	}
-
-	var searchErr error
-	idx := sort.Search(len(report.Messages), func(mid int) bool {
-		if searchErr != nil {
-			return false
-		}
-		msgs := make(map[int]struct{})
-		for i := 0; i < mid; i++ {
-			msgs[i] = struct{}{}
-		}
-		finalReport2, encodedSize2, _, err :=
-			buildSingleChainReport(ctx, lggr, hasher, tokenDataReader, encoder, report, msgs)
-		if searchErr != nil {
-			searchErr = fmt.Errorf("unable to build a single chain report (messages %d): %w", mid, err)
-		}
-
-		if (encodedSize2) <= maxSizeBytes {
-			// mid is a valid report size, try something bigger next iteration.
-			finalReport = finalReport2
-			encodedSize = encodedSize2
-			return false // not full
-		}
-		return true // full
-	})
-	if searchErr != nil {
-		return cciptypes.ExecutePluginReportSingleChain{}, 0, plugintypes.ExecutePluginCommitDataWithMessages{}, searchErr
-	}
-
-	// No messages fit into the report.
-	if idx <= 0 {
-		return cciptypes.ExecutePluginReportSingleChain{},
-			0,
-			plugintypes.ExecutePluginCommitDataWithMessages{},
-			errEmptyReport
-	}
-
-	report = markNewMessagesExecuted(finalReport, report)
-	return finalReport, encodedSize, report, nil
-}
-
-// buildSingleChainReport converts the on-chain event data stored in cciptypes.ExecutePluginCommitDataWithMessages into
-// the final on-chain report format.
-//
-// The hasher and encoding codec are provided as arguments to allow for chain-specific formats to be used.
-//
-// The messages argument indicates which messages should be included in the report. If messages is empty
-// all messages will be included. This allows the caller to create smaller reports if needed. Executed messages
-// are skipped automatically.
-func buildSingleChainReport(
-	ctx context.Context,
-	lggr logger.Logger,
-	hasher cciptypes.MessageHasher,
-	tokenDataReader TokenDataReader,
-	encoder cciptypes.ExecutePluginCodec,
-	report plugintypes.ExecutePluginCommitDataWithMessages,
-	messages map[int]struct{},
-) (cciptypes.ExecutePluginReportSingleChain, int, uint64, error) {
-	if len(messages) == 0 {
-		if messages == nil {
-			messages = make(map[int]struct{})
-		}
-		for i := 0; i < len(report.Messages); i++ {
-			messages[i] = struct{}{}
-		}
-	}
-
-	lggr.Debugw(
-		"constructing merkle tree",
-		"sourceChain", report.SourceChain,
-		"expectedRoot", report.MerkleRoot.String(),
-		"treeLeaves", len(report.Messages))
-
-	tree, err := constructMerkleTree(ctx, hasher, report)
-	if err != nil {
-		return cciptypes.ExecutePluginReportSingleChain{}, 0, 0,
-			fmt.Errorf("unable to construct merkle tree from messages for report (%s): %w", report.MerkleRoot.String(), err)
-	}
-
-	// Verify merkle root.
-	hash := tree.Root()
-	if !bytes.Equal(hash[:], report.MerkleRoot[:]) {
-		actualStr := "0x" + hex.EncodeToString(hash[:])
-		return cciptypes.ExecutePluginReportSingleChain{}, 0, 0,
-			fmt.Errorf("merkle root mismatch: expected %s, got %s", report.MerkleRoot.String(), actualStr)
-	}
-
-	// Iterate sequence range and executed messages to select messages to execute.
-	numMsgs := len(report.Messages)
-	var toExecute []int
-	var offchainTokenData [][][]byte
-	var msgInRoot []cciptypes.Message
-	executedIdx := 0
-	for i, msg := range report.Messages {
-		seqNum := report.SequenceNumberRange.Start() + cciptypes.SeqNum(i)
-		// Skip messages which are already executed
-		if executedIdx < len(report.ExecutedMessages) && report.ExecutedMessages[executedIdx] == seqNum {
-			executedIdx++
-		} else if _, ok := messages[i]; ok {
-			tokenData, err := tokenDataReader.ReadTokenData(context.Background(), report.SourceChain, msg.Header.SequenceNumber)
-			if err != nil {
-				// TODO: skip message instead of failing the whole thing.
-				//       that might mean moving the token data reading out of the loop.
-				lggr.Infow(
-					"unable to read token data",
-					"sourceChain", report.SourceChain,
-					"seqNum", msg.Header.SequenceNumber,
-					"error", err)
-				return cciptypes.ExecutePluginReportSingleChain{}, 0, 0,
-					fmt.Errorf("unable to read token data for message %d: %w", msg.Header.SequenceNumber, err)
-			}
-
-			lggr.Infow(
-				"read token data",
-				"sourceChain", report.SourceChain,
-				"seqNum", msg.Header.SequenceNumber,
-				"data", tokenData)
-			offchainTokenData = append(offchainTokenData, tokenData)
-			toExecute = append(toExecute, i)
-			msgInRoot = append(msgInRoot, msg)
-		}
-	}
-
-	lggr.Infow(
-		"selected messages from commit report for execution",
-		"sourceChain", report.SourceChain,
-		"commitRoot", report.MerkleRoot.String(),
-		"numMessages", numMsgs,
-		"toExecute", len(toExecute))
-	proof, err := tree.Prove(toExecute)
-	if err != nil {
-		return cciptypes.ExecutePluginReportSingleChain{}, 0, 0,
-			fmt.Errorf("unable to prove messages for report %s: %w", report.MerkleRoot.String(), err)
-	}
-
-	var proofsCast []cciptypes.Bytes32
-	for _, p := range proof.Hashes {
-		proofsCast = append(proofsCast, p)
-	}
-
-	finalReport := cciptypes.ExecutePluginReportSingleChain{
-		SourceChainSelector: report.SourceChain,
-		Messages:            msgInRoot,
-		OffchainTokenData:   offchainTokenData,
-		Proofs:              proofsCast,
-		ProofFlagBits:       cciptypes.BigInt{Int: slicelib.BoolsToBitFlags(proof.SourceFlags)},
-	}
-
-	// Note: ExecutePluginReport is a strict array of data, so wrapping the final report
-	//       does not add any additional overhead to the size being computed here.
-
-	// Compute the size of the encoded report.
-	encoded, err := encoder.Encode(
-		ctx,
-		cciptypes.ExecutePluginReport{
-			ChainReports: []cciptypes.ExecutePluginReportSingleChain{finalReport},
-		},
-	)
-	if err != nil {
-		lggr.Errorw("unable to encode report", "err", err, "report", finalReport)
-		return cciptypes.ExecutePluginReportSingleChain{}, 0, 0, fmt.Errorf("unable to encode report: %w", err)
-	}
-
-	gas := maxGasOverHeadGas(numMsgs, len(encoded), len(offchainTokenData))
-
-	return finalReport, len(encoded), gas, nil
-}
-
 // selectReport takes a list of reports in execution order and selects the first reports that fit within the
 // maxReportSizeBytes. Individual messages in a commit report may be skipped for various reasons, for example if an
 // out-of-order execution is detected or the message requires additional off-chain metadata which is not yet available.
@@ -455,62 +253,76 @@ func selectReport(
 	lggr logger.Logger,
 	hasher cciptypes.MessageHasher,
 	encoder cciptypes.ExecutePluginCodec,
-	tokenDataReader TokenDataReader,
-	reports []plugintypes.ExecutePluginCommitDataWithMessages,
+	tokenDataReader temp.TokenDataReader,
+	commitReports []plugintypes.ExecutePluginCommitDataWithMessages,
 	maxReportSizeBytes int,
 ) ([]cciptypes.ExecutePluginReportSingleChain, []plugintypes.ExecutePluginCommitDataWithMessages, error) {
 	// TODO: It may be desirable for this entire function to be an interface so that
 	//       different selection algorithms can be used.
 
+	builder := report.NewBuilder(ctx, lggr, hasher, tokenDataReader, encoder, uint64(maxReportSizeBytes), 99)
 	// count number of fully executed reports so that they can be removed after iterating the reports.
-	fullyExecuted := 0
-	accumulatedSize := 0
+	//fullyExecuted := 0
+	//accumulatedSize := 0
 	//accumulatedGas := uint64(0)
-	var finalReports []cciptypes.ExecutePluginReportSingleChain
-	for reportIdx, report := range reports {
+	//var finalReports []cciptypes.ExecutePluginReportSingleChain
+	for i, report := range commitReports {
 		// Reports at the end may not have messages yet.
 		if len(report.Messages) == 0 {
 			break
 		}
 
-		execReport, encodedSize, updatedReport, err :=
-			buildSingleChainReportMaxSize(ctx, lggr, hasher, tokenDataReader, encoder,
-				report,
-				maxReportSizeBytes-accumulatedSize)
-		// No messages fit into the report, stop adding more reports.
-		if errors.Is(err, errEmptyReport) {
-			break
-		}
-		if err != nil {
-			return nil, nil, fmt.Errorf("unable to build single chain report: %w", err)
-		}
-		reports[reportIdx] = updatedReport
-		accumulatedSize += encodedSize
-		finalReports = append(finalReports, execReport)
+		/*
+			execReport, encodedSize, updatedReport, err :=
+				buildSingleChainReportMaxSize(ctx, lggr, hasher, tokenDataReader, encoder,
+					report,
+					maxReportSizeBytes-accumulatedSize)
 
-		// partially executed report detected, stop adding more reports.
-		// TODO: do not break if messages were intentionally skipped.
-		if len(updatedReport.Messages) != len(updatedReport.ExecutedMessages) {
-			break
+			// No messages fit into the report, stop adding more reports.
+			if errors.Is(err, errEmptyReport) {
+				break
+			}
+			if err != nil {
+				return nil, nil, fmt.Errorf("unable to build single chain report: %w", err)
+			}
+			reports[reportIdx] = updatedReport
+			accumulatedSize += encodedSize
+			finalReports = append(finalReports, execReport)
+
+			// partially executed report detected, stop adding more reports.
+			// TODO: do not break if messages were intentionally skipped.
+			if len(updatedReport.Messages) != len(updatedReport.ExecutedMessages) {
+				break
+			}
+			fullyExecuted++
+		*/
+		var err error
+		commitReports[i], err = builder.Add(report)
+		if err != nil {
+			return nil, nil, fmt.Errorf("unable to add report to builder: %w", err)
 		}
-		fullyExecuted++
 	}
 
 	// Remove reports that are about to be executed.
-	if fullyExecuted == len(reports) {
-		reports = nil
-	} else {
-		reports = reports[fullyExecuted:]
-	}
+	/*
+		if fullyExecuted == len(commitReports) {
+			commitReports = nil
+		} else {
+			commitReports = reports[fullyExecuted:]
+		}
+	*/
 
-	lggr.Infow(
-		"selected commit reports for execution report",
-		"numReports", len(finalReports),
-		"size", accumulatedSize,
-		"incompleteReports", len(reports),
-		"maxSize", maxReportSizeBytes)
-
-	return finalReports, reports, nil
+	// TODO: Move this into build
+	/*
+		lggr.Infow(
+			"selected commit reports for execution report",
+			"numReports", len(finalReports),
+			"size", accumulatedSize,
+			"incompleteReports", len(reports),
+			"maxSize", maxReportSizeBytes)
+	*/
+	execReports, err := builder.Build()
+	return execReports, commitReports, err
 }
 
 // Outcome collects the reports from the two phases and constructs the final outcome. Part of the outcome is a fully
