+++ execute/plugin_functions.go
@@ -1,7 +1,6 @@
 package execute
 
 import (
-	"context"
 	"errors"
 	"fmt"
 	"sort"
@@ -12,8 +11,6 @@ import (
 	"github.com/smartcontractkit/libocr/commontypes"
 	"github.com/smartcontractkit/libocr/offchainreporting2plus/types"
 
-	"github.com/smartcontractkit/chainlink-common/pkg/hashutil"
-	"github.com/smartcontractkit/chainlink-common/pkg/merklemulti"
 	cciptypes "github.com/smartcontractkit/chainlink-common/pkg/types/ccipocr3"
 
 	"github.com/smartcontractkit/chainlink-ccip/execute/internal/validation"
@@ -301,59 +298,4 @@ func mergeCommitObservations(
 	}
 
 	return results, nil
-}
-
-// markNewMessagesExecuted compares an execute plugin report with the commit report metadata and marks the new messages
-// as executed.
-func markNewMessagesExecuted(
-	execReport cciptypes.ExecutePluginReportSingleChain, report plugintypes.ExecutePluginCommitDataWithMessages,
-) plugintypes.ExecutePluginCommitDataWithMessages {
-	// Mark new messages executed.
-	for i := 0; i < len(execReport.Messages); i++ {
-		report.ExecutedMessages =
-			append(report.ExecutedMessages, execReport.Messages[i].Header.SequenceNumber)
-	}
-	sort.Slice(
-		report.ExecutedMessages,
-		func(i, j int) bool { return report.ExecutedMessages[i] < report.ExecutedMessages[j] })
-
-	return report
-}
-
-// constructMerkleTree creates the merkle tree object from the messages in the report.
-func constructMerkleTree(
-	ctx context.Context,
-	hasher cciptypes.MessageHasher,
-	report plugintypes.ExecutePluginCommitDataWithMessages,
-) (*merklemulti.Tree[[32]byte], error) {
-	// Ensure we have the expected number of messages
-	numMsgs := int(report.SequenceNumberRange.End() - report.SequenceNumberRange.Start() + 1)
-	if numMsgs != len(report.Messages) {
-		return nil, fmt.Errorf(
-			"malformed report %s, unexpected number of messages: expected %d, got %d",
-			report.MerkleRoot.String(), numMsgs, len(report.Messages))
-	}
-
-	treeLeaves := make([][32]byte, 0)
-	for _, msg := range report.Messages {
-		if !report.SequenceNumberRange.Contains(msg.Header.SequenceNumber) {
-			return nil, fmt.Errorf(
-				"malformed report, message %s sequence number %d outside of report range %s",
-				report.MerkleRoot.String(), msg.Header.SequenceNumber, report.SequenceNumberRange)
-		}
-		if report.SourceChain != msg.Header.SourceChainSelector {
-			return nil, fmt.Errorf("malformed report, message %s for unexpected source chain: expected %d, got %d",
-				report.MerkleRoot.String(), report.SourceChain, msg.Header.SourceChainSelector)
-		}
-		leaf, err := hasher.Hash(ctx, msg)
-		if err != nil {
-			return nil, fmt.Errorf(
-				"unable to hash message (%d, %d): %w",
-				msg.Header.SourceChainSelector, msg.Header.SequenceNumber, err)
-		}
-		treeLeaves = append(treeLeaves, leaf)
-	}
-
-	// TODO: Do not hard code the hash function, it should be derived from the message hasher.
-	return merklemulti.NewTree(hashutil.NewKeccak(), treeLeaves)
 }
