+++ execute/report/errors.go
@@ -1,0 +1,5 @@
+package report
+
+import "errors"
+
+var ErrEmptyReport = errors.New("no messages can fit in the report")
